+++ src/protocol/types.rs
@@ -52,6 +52,6 @@ pub enum OpKind {
 #[derive(Debug, Default, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Line {
     pub text: Option<String>,
-    pub cursor: Option<Vec<u64>>,
+    pub cursor: Option<Vec<usize>>,
     pub styles: Option<Vec<i64>>,
 }
